Remove DomModule.convertPointFromNodeToPage and DomModule.convertPointFromPageToNode In order to avoid to evaluate this code in each application, so far those functions are just used by the flow. Use Point.convertPointFromPageToNode and Point.convertPointFromNodeToPage instead.
